Install the hsv3 (Tor Hidden Service V3) package on demand.
This saves up about 80mb of download and 250mb of storage for the initial
Spark installation, by not downloading the Tor Bundle immediately.

Tor will be downloaded the first time --onion is used.

In the docker environment, the bundle is downloaded into /data/spark/tor/tor-installation,
so that it persists in it the user-mounted /data volume and not re-downloaded every time.