Bug fix: don't display payments twice
C-lightning `pay` calls are idempotent and return the previous result when provided with the same hash/msatoshi/dest again.
This resulted in duplicated outgoing payments on the feed list when users pay the same invoice again.

This still affects the balance, which is incorrectly deducted by duplicates payments (until it re-syncs from `listpeers`)
and should be fixed separately.