Add npm-shrinkwrap to the list of files included in the package
Apparently, unlike the package.json file which is always included regardless of settings [0],
the npm-shrinkwrap.json file won't be included unless explicitly specified in the "files" field
(or if "files" is not set at all).

This means that npm-shrinkwrap did not ship with Spark when installed from the npm registry
before the soon-to-be-released v0.2.8. Users who installed from source or using the pre-packaged
Cordova builds (for the Android app) or the Electron builds (for the Windows, Linux and Mac desktop
apps) were not effected.

This seems like an oversight, the behaviour I would expected is to always
include the npm-shrinkwrap.json file if it exists, similarly to how the package.json
and other files (license, readme, etc) are treated. Should send a PR for this.

[0] https://docs.npmjs.com/files/package.json#files