Default the channel reserve to 0
Some channels appears not to have the `{their,our}_channel_reserve_satoshis`
fields, which was causing a type error when rendering the channel UI component.

This commit defaults the reserve to 0 when its missing, which should resolve
the issue, but is not quite right: all channels should have some reserve amount
configured for them. Seems like this might possibly be an issue with c-lightning?

The problematic channel that was encountered is in the OPENINGD state and with
this status: "Funding channel start: offered, now waiting for accept_channel"