Use `listpays` to retrieve the list of outgoing payments
So that multi-part payments are grouped together and not displayed
separately.

This requires an additional RPC call per completed outgoing payment,
until the timestamp field is added to `listpays` (expected in the next
c-lightning release).

This change also uses the `payment_hash` as the unique payment identifier
instead of the payment id, as payment ids are not available in `listpays`.

Resolves #147.