docker: Detect network subdirectory by the presence of 'lightning-rpc'
This is less prune to misdetection compared to using the database file,
but also means that mounting the parent directory will not work if spark
is started before c-lightning is.

Because of this, the documentation was updated to instruct users to
mount the network subdirectory containing the RPC file directly, rather
than relying on the automatic detection mechanism (which is more
advisable anyway).

This appears to be the underlying cause for
https://github.com/btcpayserver/btcpayserver-docker/pull/462,
due to BTCPay's unusual 'rpc-file' setting.